Do not set --std=c99 to gcc for all sources
The subproject libwebsocket can not be compiled if gcc --std=c99 is used.
On the other hand, several oter sources depend on this option.
Also some type definitions used by header files of deps/shared requiere
the macro definition _POSIX_C_SOURCE=200112L.
So this settings are made for just the sources which fail to build without
this settings.

Signed-off-by: Juergen Kosel <juergen.kosel@softing.com>