Use only 2 make jobs for build
If the machine running the build is not equipped with enough memory,
than the build process is slowed down by the swapping of memory.
To avoid this, no new jobs are started by make, if the load average
already equals the number of cores.
And for the moment the number of make jobs is set to two.

Signed-off-by: Juergen Kosel <juergen.kosel@softing.com>